requirements.txt
PyQt5>=5.15
python-vlc>=3.0